fix(clippy): resolve all mechanical lint warnings
- map_or(false, ..) → is_some_and(..) (24 instances)
- redundant closures → direct function references (8)
- format!("literal") → "literal".to_string() (6)
- manual div_ceil → .div_ceil() (6)
- needless return → implicit return (5)
- needless borrow → remove & (5)
- clone on Copy types → direct copy (5)
- iter().map(|(k,v)| v) → .values() (4)
- unnecessary casts removed (6)
- collapsible if/else/match blocks (7)
- io::Error::new(Other, e) → io::Error::other(e) (7)
- manual_ok → .ok()
- needless_range_loop → iterators
- field_reassign_with_default → struct init
- if_same_then_else → deduplicated
- unnecessary_get_then_check → contains_key

+++ blood-tools/lsp/src/semantic_tokens.rs
@@ -212,8 +212,8 @@ impl SemanticTokensProvider {
 
             self.add_token(
                 &mut tokens,
-                token.span.start_line as u32,
-                token.span.start_col as u32,
+                token.span.start_line,
+                token.span.start_col,
                 length,
                 token_type,
                 modifiers,
